+++ pyads/structs.py
@@ -341,5 +341,41 @@ class SAdsSymbolEntry(Structure):
         ("typeLength", c_uint16),
         ("commentLength", c_uint16),
         ("stringBuffer", c_ubyte * (256 * 3)),
-        # 3 strings contained, with max length 256
+        # 3 strings contained, with max length 256 each
+    ]
+
+    def _get_string(self, offset, length):
+        return bytes(self.stringBuffer[offset : offset + length]).decode("utf-8")
+
+    @property
+    def name(self):
+        "The symbol name"
+        return self._get_string(0, self.nameLength)
+
+    @property
+    def type_name(self):
+        "The qualified type name, including the namespace"
+        return self._get_string(self.nameLength + 1, self.typeLength)
+
+    @property
+    def comment(self):
+        "User-defined comment"
+        return self._get_string(
+            self.nameLength + self.typeLength + 2, self.commentLength
+        )
+
+
+class SAdsSumRequest(Structure):
+    """ADS sum request structure.
+    
+    :ivar iGroup: indexGroup of request
+    :ivar iOffs: indexOffset of request
+    :ivar size: size of request
+    """
+
+    _pack_ = 1
+    _fields_ = [
+        ("iGroup", c_uint32),
+        ("iOffset", c_uint32),
+        ("size", c_uint32),
     ]